requirements.txt
boto3boto3
botocore
uuid
typing-extensions
